Moves aggregator annotations to XML
