+++ src/Resources/config/doctrine/aggregator.mongodb.xml
@@ -4,9 +4,9 @@
                         xsi:schemaLocation="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping
                         http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping.xsd">
 
-    <mapped-superclass name="Algolia\SearchBundle\Aggregator">
-        <field name="objectID" column="id" type="string">
+    <mapped-superclass name="Algolia\SearchBundle\Entity\Aggregator">
+        <id name="objectID" type="string">
              <generator strategy="AUTO"/>
-        </field>
+        </id>
     </mapped-superclass>
 </doctrine-mongo-mapping>
