+++ src/Resources/config/doctrine/aggregator.orm.xml
@@ -4,9 +4,9 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                   http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <mapped-superclass name="Algolia\SearchBundle\Aggregator">
-        <field name="objectID" column="id" type="string">
-             <generator strategy="AUTO"/>
-        </field>
+    <mapped-superclass name="Algolia\SearchBundle\Entity\Aggregator">
+        <id name="objectID" type="string">
+            <generator strategy="AUTO"/>
+        </id>
     </mapped-superclass>
 </doctrine-mapping>
