+++ tests/TestApp/Entity/Image.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Algolia\SearchBundle\Entity;
+namespace Algolia\SearchBundle\TestApp\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
